feat: Add ILazyLoad component to Ice page
This commit adds the ILazyLoad component to the Ice page in order to lazy load the image. The ILazyLoad component is imported from the LazyLoad module and wraps the image tag. This improves the performance of the page by only loading the image when it becomes visible in the viewport.

Refactor the code to use ILazyLoad component instead of the commented out LazyLoad component.